added more csrf

+++ openeclass/modules/user/user.php
@@ -76,27 +76,27 @@ $tool_content = "";
 if ($is_adminOfCourse) {
 
         // Handle user removal / status change
-        if (isset($_GET['giveAdmin'])) {
+        if (isset($_GET['giveAdmin']) && (isset($_GET['_token']) && ($_GET['_token'] == $_SESSION['_token']))) {
                 $new_admin_gid = intval($_GET['giveAdmin']);
                 db_query("UPDATE cours_user SET statut = 1
                                 WHERE user_id = $new_admin_gid AND cours_id = $cours_id", $mysqlMainDb);
-        } elseif (isset($_GET['giveTutor'])) {
+        } elseif (isset($_GET['giveTutor']) && (isset($_GET['_token']) && ($_GET['_token'] == $_SESSION['_token']))) {
                 $new_tutor_gid = intval($_GET['giveTutor']);
                 db_query("UPDATE cours_user SET tutor = 1
                                 WHERE user_id = $new_tutor_gid AND cours_id = $cours_id", $mysqlMainDb);
                 db_query("DELETE FROM user_group WHERE user = $new_tutor_gid", $currentCourseID);
-        } elseif (isset($_GET['removeAdmin'])) {
+        } elseif (isset($_GET['removeAdmin']) && (isset($_GET['_token']) && ($_GET['_token'] == $_SESSION['_token']))) {
                 $removed_admin_gid = intval($_GET['removeAdmin']);
                 db_query("UPDATE cours_user SET statut = 5
                                 WHERE user_id <> $uid AND
                                       user_id = $removed_admin_gid AND
                                       cours_id = $cours_id", $mysqlMainDb);
-        } elseif (isset($_GET['removeTutor'])) {
+        } elseif (isset($_GET['removeTutor']) && (isset($_GET['_token']) && ($_GET['_token'] == $_SESSION['_token']))) {
                 $removed_tutor_gid = intval($_GET['removeTutor']);
                 db_query("UPDATE cours_user SET tutor = 0
                                 WHERE user_id = $removed_tutor_gid AND
                                       cours_id = $cours_id", $mysqlMainDb);
-        } elseif (isset($_GET['unregister'])) {
+        } elseif (isset($_GET['unregister']) &&(isset($_GET['_token']) && ($_GET['_token'] == $_SESSION['_token']))) {
                 $unregister_gid = intval($_GET['unregister']);
                 $unregister_ok = true;
                 // Security: don't remove myself except if there is another prof
@@ -301,17 +301,35 @@ while ($myrow = mysql_fetch_array($result)) {
         if(isset($status) && ($status["$currentCourseID"]=='1' OR $status["$currentCourseID"]=='2')) {
                 // tutor right
                 if ($myrow['tutor'] == '0') {
-                        $tool_content .= "<td valign='top' align='center' class='add_user'><a href='$_SERVER[PHP_SELF]?giveTutor=$myrow[user_id]' title='$langGiveTutor'>$langAdd</a></td>";
+                        $tool_content .="<form id='my_form".$myrow[user_id]."tutor' method=\"get\" action=\"http://localhost:8001/modules/user/user.php\">";
+                        $tool_content .="<input type='hidden' name='giveTutor' value='$myrow[user_id]'/>";
+                        $tool_content .="<input type='hidden' name='_token' value='".$_SESSION['_token']."'/>";
+                        $tool_content .= "<td valign='top' align='center' class='add_user'><a href='javascript:{}' onclick=\"document.getElementById('my_form".$myrow[user_id]."tutor').submit(); return false;\" title='$langGiveTutor'>$langAdd</a></td>";
+                        $tool_content .= "</form>";
                 } else {
-                        $tool_content .= "<td class='highlight' align='center'>$langTutor<br /><a href='$_SERVER[PHP_SELF]?removeTutor=$myrow[user_id]' title='$langRemoveRight'>$langRemove</a></td>";
+                    $tool_content .="<form id='my_form".$myrow[user_id]."tutor' method=\"get\" action=\"http://localhost:8001/modules/user/user.php\">";
+                    $tool_content .="<input type='hidden' name='removeTutor' value='$myrow[user_id]'/>";
+                    $tool_content .="<input type='hidden' name='_token' value='".$_SESSION['_token']."'/>";
+                        $tool_content .= "<td class='highlight' align='center'>$langTutor<br /><a href='javascript:{}' onclick=\"document.getElementById('my_form".$myrow[user_id]."tutor').submit(); return false;\" title='$langRemoveRight'>$langRemove</a></td>";
+                    $tool_content .= "</form>";
+
                 }
 
                 // admin right
                 if ($myrow['user_id'] != $_SESSION["uid"]) {
                         if ($myrow['statut']=='1') {
-                                $tool_content .= "<td class='highlight' align='center'>$langAdministrator<br /><a href='$_SERVER[PHP_SELF]?removeAdmin=$myrow[user_id]' title='$langRemoveRight'>$langRemove</a></td>";
+                            $tool_content .="<form id='my_form".$myrow[user_id]."admin' method=\"get\" action=\"http://localhost:8001/modules/user/user.php\">";
+                            $tool_content .="<input type='hidden' name='removeAdmin' value='$myrow[user_id]'/>";
+                            $tool_content .="<input type='hidden' name='_token' value='".$_SESSION['_token']."'/>";
+                                $tool_content .= "<td class='highlight' align='center'>$langAdministrator<br /><a href='javascript:{}' onclick=\"document.getElementById('my_form".$myrow[user_id]."admin').submit(); return false;\" title='$langRemoveRight'>$langRemove</a></td>";
+                            $tool_content .= "</form>";
+
                         } else {
-                                $tool_content .= "<td valign='top' align='center' class='add_user'><a href='$_SERVER[PHP_SELF]?giveAdmin=$myrow[user_id]' title='$langGiveAdmin'>$langAdd</a></td>";
+                            $tool_content .="<form id='my_form".$myrow[user_id]."admin' method=\"get\" action=\"http://localhost:8001/modules/user/user.php\">";
+                            $tool_content .="<input type='hidden' name='giveAdmin' value='$myrow[user_id]'/>";
+                            $tool_content .="<input type='hidden' name='_token' value='".$_SESSION['_token']."'/>";
+                                $tool_content .= "<td valign='top' align='center' class='add_user'><a href='javascript:{}' onclick=\"document.getElementById('my_form".$myrow[user_id]."admin').submit(); return false;\" title='$langGiveAdmin'>$langAdd</a></td>";
+                            $tool_content .= "</form>";
                         }
                 } else {
                         if ($myrow['statut']=='1') {
@@ -322,7 +340,11 @@ while ($myrow = mysql_fetch_array($result)) {
                 }
                 $tool_content .= "<td valign='top' align='center'>";
                 $alert_uname = $myrow['prenom'] . " " . $myrow['nom'];
-                $tool_content .= "<a href='$_SERVER[PHP_SELF]?unregister=$myrow[user_id]' onClick=\"return confirmation('".mysql_real_escape_string($alert_uname)."');\"><img src='../../template/classic/img/delete.gif' title='$langDelete' /></a>";
+            $tool_content .="<form id='my_form".$myrow[user_id]."unregister' method=\"get\" action=\"http://localhost:8001/modules/user/user.php\">";
+            $tool_content .="<input type='hidden' name='unregister' value='$myrow[user_id]'/>";
+            $tool_content .="<input type='hidden' name='_token' value='".$_SESSION['_token']."'/>";
+                $tool_content .= "<a href='javascript:{}' onclick=\"document.getElementById('my_form".$myrow[user_id]."unregister').submit(); return false;\"> <img src='../../template/classic/img/delete.gif' title='$langDelete' /></a>";
+                $tool_content .="</form>";
         }	// admin only
         $tool_content .= "</td></tr>";$i++;
 } 	// end of while
